extract shared AgentHistoryDrawer + dock both history drawers left
The bottom-input history drawer and the in-overlay chat history
drawer were two separate implementations and they diverged:
delete only worked in the overlay version, and one slid in from
the left while the other slid in from the right. Pulled them
together into a shared AgentHistoryDrawer component, both keyed
on `side="left"`.

The shared component owns the conversation list (refetches on
every open), the loading state, and the delete + ConfirmOverlay
flow. AgentChat passes the active conversation id (for row
highlighting), a `newChat` callback (for the optional New chat
row at the top), and an `onDeleted` callback so it can bounce
out of the welcome screen if the user deletes the conversation
they're currently viewing. BottomAgentInput just passes
`onSelect` and forgets about all the drawer plumbing.

Removes from AgentChat: allConversations state, the second list
fetch in the load effect, refreshConversationList, the
post-stream refresh, pendingDeleteId/deleting state,
confirmDelete, the inline ConversationRow component, and the
now-unused formatRelative helper.
